pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "dorkforge"
version = "1.0.0"
description = "Professional Google Dork Generator with AI & Web UI"
authors = ["Ayberk <ayberk@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/mhmtayberk/DorkForge"
repository = "https://github.com/mhmtayberk/DorkForge"
keywords = ["security", "pentesting", "osint", "google-dorks", "reconnaissance"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-cors = "^4.0.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pyyaml = "^6.0"
rich = "^13.0.0"
questionary = "^2.0.0"
pydantic = "^2.0.0"
google-generativeai = "^0.8.0"
openai = "^1.1.0"
anthropic = "^0.18.0"
groq = "^0.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-cov = "^4.1"
pytest-mock = "^3.11"
mypy = "^1.5"
ruff = "^0.1"
black = "^23.0"
pre-commit = "^3.4"
types-pyyaml = "^6.0"

[tool.poetry.scripts]
dorkforge = "app:app"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
